Move command to a different file
